+++ packages/ast-spec/src/legacy-fixtures/basics/fixtures/class-with-private-optional-property/fixture.ts
@@ -1,0 +1,7 @@
+// TODO: This fixture might be too large, and if so should be split up.
+
+class Foo {
+  #prop?: string;
+  #propExplicitWithValue?: string = '';
+  #propImplicitWithValue? = '';
+}