+++ packages/ast-spec/src/legacy-fixtures/basics/fixtures/class-with-property-values/fixture.ts
@@ -1,0 +1,9 @@
+// TODO: This fixture might be too large, and if so should be split up.
+
+class Foo {
+  a = 2;
+  b = {};
+  c = [];
+  d = '';
+  e = new Array([{}, [], 2]);
+}