+++ packages/ast-spec/src/legacy-fixtures/basics/fixtures/interface-without-type-annotation/fixture.ts
@@ -1,0 +1,5 @@
+// TODO: This fixture might be too large, and if so should be split up.
+
+interface test {
+  foo;
+}