+++ packages/shared-fixtures/fixtures/javascript/classes/class-private-identifier-method.src.js
@@ -1,7 +1,0 @@
-class Foo {
-  #bar() {}
-
-  constructor() {
-    this.#bar();
-  }
-}