+++ packages/shared-fixtures/fixtures/javascript/templateStrings/multi-line-template-string.src.js
@@ -1,6 +1,0 @@
-`The last man on Earth
-                 sat alone in a room.
-    There was
-    a knock
-    on the
-    door...`;