+++ packages/shared-fixtures/fixtures/jsx/tag-names-with-multi-dots-multi.src.js
@@ -1,3 +1,0 @@
-<a.b.c.d.e.f>
-  <this.const.declare.static></this.const.declare.static>
-</a.b.c.d.e.f>;