+++ packages/shared-fixtures/fixtures/typescript/basics/class-with-constructor-and-parameter-proptery-with-override-modifier.src.ts
@@ -1,5 +1,0 @@
-class SpecializedComponent extends SomeComponent {
-  constructor(override param: string) {
-      super();
-  }
-}