+++ packages/shared-fixtures/fixtures/typescript/basics/import-export-equal-type-declaration.src.ts
@@ -1,1 +1,0 @@
-export import type foo = require('bar');