+++ packages/shared-fixtures/fixtures/typescript/basics/import-type-star-as-ns.src.ts
@@ -1,1 +1,0 @@
-import type * as foo from './bar';