+++ packages/shared-fixtures/fixtures/typescript/basics/type-assertion-in-function.src.ts
@@ -1,3 +1,0 @@
-function assertsString(x: any): asserts x {
-    return
-}