+++ packages/shared-fixtures/fixtures/typescript/basics/type-assertion-with-guard-in-function.src.ts
@@ -1,3 +1,0 @@
-function assertsStringGuard(x: any): asserts x is string {
-    return
-}