+++ packages/shared-fixtures/fixtures/typescript/errorRecovery/interface-property-static.src.ts
@@ -1,4 +1,0 @@
-interface Foo {
-  static a: string;
-}
-