+++ packages/shared-fixtures/fixtures/typescript/types/this-type.src.ts
@@ -1,5 +1,0 @@
-class Message {
-  clone(): this {
-    return this;
-  }
-}